ingestion/requirements.txt
psycopg2-binary
python-dotenv
cryptography
amazon-orders